move debug prints to if __debug__:
This change is not a no-op because it
means the default way of running python will
print the debug statements.  The shebang is adjusted
to be correct, however people who do python scriptname.py
will now see the debug output.

On the other hand, if you WANT to see the debug output you
no longer have to go through the source code and uncomment
a bunch of things.

Also, this test for __debug__ is optimized out so it
doesn't make add overhead to the script. (like
adding a homerolled DEBUG global would do)